dom/animation/test/css-transitions/test_animation-pausing.html

<!doctype html>
<meta charset=utf-8>
<title>Pausing a CSSTransition</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<div id="log"></div>
<script>
'use strict';

promise_test(async t => {
  const div = addDiv(t);
  const cs = getComputedStyle(div);

  div.style.marginLeft = '0px';
  cs.marginLeft; // Flush style to set up transition start point
  div.style.transition = 'margin-left 100s';
  div.style.marginLeft = '10000px';
  cs.marginLeft;

  const animation = div.getAnimations()[0];
  let previousProgress = animation.effect.getComputedTiming().progress;
  assert_equals(previousProgress, 0, 'Initial value of progress is zero');

  await animation.ready;
  await waitForNextFrame();

  assert_greater_than(animation.effect.getComputedTiming().progress,
                      previousProgress,
                      'Iteration progress is initially increasing');
  animation.pause();
  await animation.ready;

  previousProgress = animation.effect.getComputedTiming().progress;
  await waitForNextFrame();

  assert_equals(animation.effect.getComputedTiming().progress,
                previousProgress,
                'Iteration progress does not increase after calling pause()');
  previousProgress = animation.effect.getComputedTiming().progress;
  animation.play();
  await animation.ready;
  await waitForNextFrame();

  assert_greater_than(animation.effect.getComputedTiming().progress,
                      previousProgress,
                      'Iteration progress increases after calling play()');
}, 'pause() and play() a transition');

</script>
